% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{generate_single_report}
\alias{generate_single_report}
\title{Generate Single Report}
\usage{
generate_single_report(
  catalogue,
  bootstraps,
  bootstraps_experimental,
  similarity,
  tally,
  dataset,
  parquet_path = NULL,
  sample_information = NULL,
  outdir
)
}
\arguments{
\item{catalogue}{.expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{.bootstrap_summary file path which contains the optimal bootstrap statistics for each signature}

\item{tally}{.tally file path which contain the decomposition of mutations in the sample}

\item{dataset}{the COSMIC signature dataset being used}

\item{parquet_path}{a path to the folder of a parquet file which describes the signature models fitted to each sample in the database (optional parameter)}

\item{sample_information}{a .metadata.tsv. file which contains sample information (optional parameter)}

\item{outdir}{the output directory you want to save reports in}

\item{bootstraps_experimenal}{.expo_bootstraps file path which contain the experimental bootstrap statistics for each signature}
}
\value{
full signature html report
}
\description{
This function auto-generates the signature report.
}
