% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{generate_summary_layer}
\alias{generate_summary_layer}
\title{Generate Summary Layer}
\usage{
generate_summary_layer(
  outdir,
  catalogue,
  bootstraps,
  tally,
  dataset,
  catalogue2,
  bootstraps2,
  tally2,
  dataset2,
  catalogue3,
  bootstraps3,
  tally3,
  dataset3
)
}
\arguments{
\item{outdir}{the output directory you want to save reports in}

\item{catalogue}{SBS96 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{SBS96 .bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{tally}{SBS96 .tally file path which contain the decompositions of the mutations}

\item{dataset}{the COSMIC signature dataset being used for SBS96}

\item{catalogue2}{DBS78 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps2}{DBS78 .bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{tally2}{DBS78 .tally file path which contain the decompositions of the mutations}

\item{dataset2}{the COSMIC signature dataset being used for DBS78}

\item{catalogue3}{ID83 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps3}{ID83 .bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{tally3}{ID83 .tally file path which contain the decompositions of the mutations}

\item{dataset3}{the COSMIC signature dataset being used for ID83}

\item{dimensionality_reduction}{the dimensionality reduction of the samples in the database}
}
\value{
full signature html reports
}
\description{
This function auto-generates a summary layer. Note: Please input
the files in this order: (SBS96.expo file, SBS96.bootstrap file, SBS96.tally file, SBS96.dataset,
DBS78.expo file, DBS78.bootstrap file, DBS78.tally file, DBS78.dataset, ID83.expo file,
ID83.bootstrap file, ID83.tally file, ID83.dataset).
}
