\name{sigstory-package}
\alias{sigstory-package}
\alias{sigstory}
\docType{package}
\title{
\packageTitle{sigstory}
}
\description{
\packageDescription{sigstory}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{sigstory}

Maintainer: \packageMaintainer{sigstory}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = sigstory)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{sigstory}}
%\section{Documentation Index}{\packageIndices{sigstory}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
