% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{sigstory}
\alias{sigstory}
\title{Sigstory}
\usage{
sigstory(
  outidr,
  catalogue,
  bootstraps,
  bootstraps_experimental,
  similarity,
  tally,
  dataset,
  parquet_path = NULL,
  catalogue2 = NULL,
  bootstraps2 = NULL,
  bootstraps_experimental2 = NULL,
  similarity2 = NULL,
  tally2 = NULL,
  dataset2 = NULL,
  parquet_path2 = NULL,
  catalogue3 = NULL,
  bootstraps3 = NULL,
  bootstraps_experimental3 = NULL,
  similarity3 = NULL,
  tally3 = NULL,
  dataset3 = NULL,
  parquet_path3 = NULL,
  sample_information = NULL
)
}
\arguments{
\item{catalogue}{SBS96 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{SBS96 .bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{similarity}{SBS96 similarity file path which contains cosine similarity to other samples in. database}

\item{tally}{SBS96 .tally file path which contain the decompositions of the mutations}

\item{dataset}{the COSMIC signature dataset being used for SBS96}

\item{catalogue2}{DBS78 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps2}{DBS78 .bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{similarity2}{DBS78 .similarity file path which contains cosine similarity to other samples in. database}

\item{tally2}{DBS78 .tally file path which contain the decompositions of the mutations}

\item{dataset2}{the COSMIC signature dataset being used for DBS78}

\item{catalogue3}{ID83 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps3}{ID83 .bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{similarity3}{ID83 .similarity file path which contains cosine similarity to other samples in. database}

\item{tally3}{ID83 .tally file path which contain the decompositions of the mutations}

\item{dataset3}{the COSMIC signature dataset being used for ID83}

\item{sample_information}{a .metadata.tsv. file which contains sample information}

\item{outdir}{the output directory you want to save reports in}

\item{paruqet_path}{the path to a paruqet database for SBS96}

\item{paruqet_path2}{the path to a paruqet database for DBS78}

\item{paruqet_path3}{the path to a paruqet database for ID83}
}
\value{
full signature html reports
}
\description{
This function auto-generates signature reports and a summary layer. Note: Please input
the files in this order: (SBS96.expo file, SBS96.bootstrap file, SBS96.tally file, SBS96.dataset, SBS96 parquet path,
DBS78.expo file, DBS78.bootstrap file, DBS78.tally file, DBS78.dataset, DBS78 parquet path, ID83.expo file,
ID83.bootstrap file, ID83.tally file, ID83.dataset, ID83 parquet path).
}
