% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{generate_single_report}
\alias{generate_single_report}
\title{Generate Single Report}
\usage{
generate_single_report(
  outdir,
  exposures,
  bootstraps,
  bootstraps_experimental,
  tally,
  dataset,
  similarity = NULL,
  dimensionality_reduction = NULL,
  parquet_path = NULL,
  sample_information = NULL
)
}
\arguments{
\item{outdir}{the output directory you want to save reports in}

\item{exposures}{.expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{.bootstrap_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental}{.expo_bootstraps file path which contains the experimental bootstrap statistics for each signature}

\item{tally}{.tally file path which contains the decomposition of mutations in the sample}

\item{dataset}{the COSMIC signature dataset being used}

\item{similarity}{.similarity file path which contains cosine similarity to other samples in the database (optional parameter)}

\item{dimensionality_reduction}{the dimensionality reduction .csv file of the samples in the database (optional parameter)}

\item{parquet_path}{a path to the folder of a parquet file which describes the signature models fitted to each sample in the database (optional parameter)}

\item{sample_information}{a .metadata.tsv file which contains sample information (optional parameter)}
}
\value{
Mutational Signature Report (HTML File)
}
\description{
This function auto-generates a full mutational signature report.
}
