% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{generate_summary_layer}
\alias{generate_summary_layer}
\title{Generate Summary Layer}
\usage{
generate_summary_layer(
  outdir,
  exposures,
  bootstraps,
  tally,
  dataset,
  exposures2,
  bootstraps2,
  tally2,
  dataset2,
  exposures3,
  bootstraps3,
  tally3,
  dataset3,
  sample_information = NULL,
  dimensionality_reduction_overall = NULL
)
}
\arguments{
\item{outdir}{the output directory you want to save reports in}

\item{exposures}{SBS96 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{SBS96 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{tally}{SBS96 .tally file path which contain the decompositions of the mutations}

\item{dataset}{the COSMIC signature dataset being used for SBS96}

\item{exposures2}{DBS78 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps2}{DBS78 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{tally2}{DBS78 .tally file path which contain the decompositions of the mutations}

\item{dataset2}{the COSMIC signature dataset being used for DBS78}

\item{exposures3}{ID83 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps3}{ID83 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{tally3}{ID83 .tally file path which contain the decompositions of the mutations}

\item{dataset3}{the COSMIC signature dataset being used for ID83}

\item{sample_information}{a .metadata.tsv. file which contains sample information}

\item{dimensionality_reduction_overall}{the dimensionality reduction .csv file of all samples in the database across all mutation types}
}
\value{
Mutational Signature Summary Layer for all mutation types (HTML File)
}
\description{
This function auto-generates a summary layer. Note: Please input the files in this order:
\itemize{
\item outidr (outdirectory path)
\item SBS96 .expo file
\item SBS96 .bootstrap_summary file
\item SBS96 .tally file
\item SBS96 COSMIC dataset
\item DBS78 .expo file
\item DBS78 .bootstrap file
\item DBS78 .tally file
\item DBS78 COSMIC dataset
\item ID83 .expo file
\item ID83 .bootstrap file
\item ID83 .tally file
\item ID83  COSMIC dataset
\item sample_information .csv file (optional)
\item dimensionality reduction .csv file for all mutation types (optional)
}
}
