% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{sigstory}
\alias{sigstory}
\title{Sigstory}
\usage{
sigstory(
  outidr,
  exposures,
  bootstraps,
  bootstraps_experimental,
  tally,
  dataset,
  similarity = NULL,
  dimensionality_reduction = NULL,
  parquet_path = NULL,
  exposures2 = NULL,
  bootstraps2 = NULL,
  bootstraps_experimental2 = NULL,
  tally2 = NULL,
  dataset2 = NULL,
  similarity2 = NULL,
  dimensionality_reduction2 = NULL,
  parquet_path2 = NULL,
  exposures3 = NULL,
  bootstraps3 = NULL,
  bootstraps_experimental3 = NULL,
  tally3 = NULL,
  dataset3 = NULL,
  similarity3 = NULL,
  dimensionality_reduction3 = NULL,
  parquet_path3 = NULL,
  sample_information = NULL,
  dimensionality_reduction_overall = NULL
)
}
\arguments{
\item{exposures}{SBS96 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{SBS96 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental}{SBS96 .expo_bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{tally}{SBS96 .tally file path which contains the decompositions of the mutations}

\item{dataset}{the COSMIC signature dataset being used for SBS96}

\item{similarity}{SBS96 similarity file path which contains cosine similarity to other samples in the database}

\item{dimensionality_reduction}{the dimensionality reduction .csv file of the samples in the database for SBS96}

\item{parquet_path}{the path to a parquet database for SBS96}

\item{exposures2}{DBS78 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps2}{DBS78 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental2}{DBS78 .expo_bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{tally2}{DBS78 .tally file path which contains the decompositions of the mutations}

\item{dataset2}{the COSMIC signature dataset being used for DBS78}

\item{similarity2}{DBS78 .similarity file path which contains cosine similarity to other samples in the database}

\item{dimensionality_reduction2}{the dimensionality reduction .csv file of the samples in the database for DBS78}

\item{parquet_path2}{the path to a parquet database for DBS78}

\item{exposures3}{ID83 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps3}{ID83 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental3}{ID83 .expo_bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{tally3}{ID83 .tally file path which contains the decompositions of the mutations}

\item{dataset3}{the COSMIC signature dataset being used for ID83}

\item{similarity3}{ID83 .similarity file path which contains cosine similarity to other samples in the database}

\item{dimensionality_reduction3}{the dimensionality reduction .csv file of the samples in the database for ID83}

\item{parquet_path3}{the path to a parquet database for ID83}

\item{sample_information}{a .metadata.tsv file which contains sample information}

\item{dimensionality_reduction_overall}{the dimensionality reduction .csv file of the samples in the database across all types}

\item{outdir}{the output directory you want to save reports in}
}
\value{
Mutational Signature Report (If NULL values exist for files) or Mutational Signature Reports for each mutation type + Mutational Signature Summary Layer (HTML Reports)
}
\description{
This function auto-generates signature reports and a summary layer.
Note: Please input the files in this order:
\itemize{
\item outdir (outdirectory path)
\item SBS96 .expo file
\item SBS96 .bootstrap_summary file
\item SBS96 .expo_bootstrap file
\item SBS96 .tally file
\item SBS96 .similarity file
\item SBS96 COSMIC dataset
\item SBS96 dimensionality reduction .csv file
\item SBS96 parquet path (optional)
\item DBS78 .expo file
\item DBS78 .bootstrap file
\item DBS78 .expo_bootstrap file
\item DBS78 .tally file
\item DBS78 .similarity file
\item DBS78 COSMIC dataset
\item DBS78 dimensionality reduction .csv file
\item DBS78 parquet path (optional)
\item ID83 .expo file
\item ID83 .bootstrap file
\item ID83 .expo_bootstrap file
\item ID83 .tally file
\item ID83 .similarity file
\item ID83  COSMIC dataset
\item ID83 dimensionality reduction .csv file
\item ID83 parquet path  (optional)
\item sample_information .csv file (optional)
\item dimensionality reduction .csv file for all mutation types (optional)
}
}
\details{
This function expects an outdirectory name and at least 1 of each:
\itemize{
\item .expo file
\item .bootstraps_summary file
\item .expo_bootstraps file
\item .similiarity file
\item .tally file
\item COSMIC Dataset
\item dimensionality reduction .csv file
}

If only one of the above is provided then only a single HTML report will be created. This function will also only
produce a singular HTML report if a NULL is entered for any of the above files (if more than one are inputted). To produce
a summary layer, all files must be produced (from above) and for additional information in the reports you may want to add a:
\itemize{
\item parquet path
\item overall sample information .csv file
\item overall dimensionality reduction .csv file
}
}
