% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigstory_functions.R
\name{sigstory}
\alias{sigstory}
\title{Sigstory}
\usage{
sigstory(
  outidr,
  exposures,
  bootstraps,
  bootstraps_experimental,
  similarity,
  tally,
  dataset,
  dimensionality_reduction,
  parquet_path = NULL,
  exposures2 = NULL,
  bootstraps2 = NULL,
  bootstraps_experimental2 = NULL,
  similarity2 = NULL,
  tally2 = NULL,
  dataset2 = NULL,
  dimensionality_reduction2 = NULL,
  parquet_path2 = NULL,
  exposures3 = NULL,
  bootstraps3 = NULL,
  bootstraps_experimental3 = NULL,
  similarity3 = NULL,
  tally3 = NULL,
  dataset3 = NULL,
  dimensionality_reduction3 = NULL,
  parquet_path3 = NULL,
  sample_information = NULL,
  dimensionality_reduction_overall = NULL
)
}
\arguments{
\item{exposures}{SBS96 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps}{SBS96 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental}{SBS96 .expo_bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{similarity}{SBS96 similarity file path which contains cosine similarity to other samples in. database}

\item{tally}{SBS96 .tally file path which contain the decompositions of the mutations}

\item{dataset}{the COSMIC signature dataset being used for SBS96}

\item{dimensionality_reduction}{the dimensionality reduction .csv file of the samples in the database for SBS96}

\item{parquet_path}{the path to a parquet database for SBS96}

\item{exposures2}{DBS78 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps2}{DBS78 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental2}{DBS78 .expo_bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{similarity2}{DBS78 .similarity file path which contains cosine similarity to other samples in. database}

\item{tally2}{DBS78 .tally file path which contain the decompositions of the mutations}

\item{dataset2}{the COSMIC signature dataset being used for DBS78}

\item{dimensionality_reduction2}{the dimensionality reduction .csv file of the samples in the database for DBS78}

\item{parquet_path2}{the path to a parquet database for DBS78}

\item{exposures3}{ID83 .expo file path which contains the optimal contributions of signatures in the sample}

\item{bootstraps3}{ID83 .bootstraps_summary file path which contains the optimal bootstrap statistics for each signature}

\item{bootstraps_experimental3}{ID83 .expo_bootstraps file path which contains the optimal bootstrap statistics for each signature}

\item{similarity3}{ID83 .similarity file path which contains cosine similarity to other samples in. database}

\item{tally3}{ID83 .tally file path which contain the decompositions of the mutations}

\item{dataset3}{the COSMIC signature dataset being used for ID83}

\item{dimensionality_reduction3}{the dimensionality reduction .csv file of the samples in the database for ID83}

\item{parquet_path3}{the path to a parquet database for ID83}

\item{sample_information}{a .metadata.tsv. file which contains sample information}

\item{dimensionality_reduction_overall}{the dimensionality reduction .csv file of the samples in the database across all types}

\item{outdir}{the output directory you want to save reports in}
}
\value{
full signature html reports
}
\description{
This function auto-generates signature reports and a summary layer. Note: Please input
the files in this order:
}
\details{
(outidr (outdirectory path), SBS96.expo file, SBS96.bootstrap_summary file, SBS96 .expo_bootstrap file, SBS96 .tally file, SBS96 .similarity file, SBS96.dataset, SBS96 dimensionality reduction .csv file, SBS96 parquet path,
DBS78.expo file, DBS78.bootstrap file, DBS78 .expo_bootstrap file, DBS78.tally file, DBS78 .similarity file, DBS78.dataset, DBS78 dimensionality reduction .csv file, DBS78 parquet path,
ID83.expo file, ID83.bootstrap file, ID83 .expo_bootstrap file, ID83.tally file, ID83 .similarity file, ID83.dataset, ID83 dimensionality reduction .csv file, ID83 parquet path,
sample_information .csv file, dimensionality reduction .csv file for all mutation types).
}
